Refactor : common폴더 uitl로 이름 변경
